Add link flow.
