+++ crates/core/src/flow/mod.rs
@@ -1,4 +1,5 @@
 pub mod install;
+pub mod link;
 pub mod locate;
 pub mod resolve;
 pub mod setup;
