+++ crates/core/src/tool.rs
@@ -2,19 +2,14 @@ use crate::error::ProtoError;
 use crate::helpers::get_proto_version;
 use crate::layout::{Inventory, Product};
 use crate::proto::ProtoEnvironment;
-use crate::shim_registry::{Shim, ShimRegistry, ShimsMap};
-use miette::IntoDiagnostic;
 use proto_pdk_api::*;
-use proto_shim::*;
 use rustc_hash::{FxHashMap, FxHashSet};
 use starbase_styles::color;
-use starbase_utils::fs;
-use std::collections::BTreeMap;
-use std::fmt::{self, Debug};
+use std::fmt;
 use std::path::{Path, PathBuf};
 use std::sync::Arc;
 use std::time::Duration;
-use tracing::{debug, instrument, warn};
+use tracing::{debug, instrument};
 use warpgate::{
     host::{create_host_functions, HostData},
     Id, PluginContainer, PluginLocator, PluginManifest, VirtualPath, Wasm,
@@ -75,7 +70,7 @@ impl Tool {
     }
 
     #[instrument(name = "new_tool", skip(proto, wasm))]
-    pub async fn load<I: AsRef<Id> + Debug, P: AsRef<ProtoEnvironment>>(
+    pub async fn load<I: AsRef<Id> + fmt::Debug, P: AsRef<ProtoEnvironment>>(
         id: I,
         proto: P,
         wasm: Wasm,
@@ -293,187 +288,6 @@ impl Tool {
 
         if modified {
             manifest.save()?;
-        }
-
-        Ok(())
-    }
-}
-
-// BINARIES, SHIMS
-
-impl Tool {
-    /// Create all executables for the current tool.
-    /// - Locate the primary binary to execute.
-    /// - Generate shims to `~/.proto/shims`.
-    /// - Symlink bins to `~/.proto/bin`.
-    #[instrument(skip(self))]
-    pub async fn create_executables(
-        &mut self,
-        force_shims: bool,
-        force_bins: bool,
-    ) -> miette::Result<()> {
-        self.locate_exe_file().await?;
-        self.generate_shims(force_shims).await?;
-        self.symlink_bins(force_bins).await?;
-
-        Ok(())
-    }
-
-    /// Create shim files for the current tool if they are missing or out of date.
-    /// If find only is enabled, will only check if they exist, and not create.
-    #[instrument(skip(self))]
-    pub async fn generate_shims(&mut self, force: bool) -> miette::Result<()> {
-        let shims = self.resolve_shim_locations().await?;
-
-        if shims.is_empty() {
-            return Ok(());
-        }
-
-        let is_outdated = self.inventory.manifest.shim_version != SHIM_VERSION;
-        let force_create = force || is_outdated;
-        let find_only = !force_create;
-
-        if force_create {
-            debug!(
-                tool = self.id.as_str(),
-                shims_dir = ?self.proto.store.shims_dir,
-                shim_version = SHIM_VERSION,
-                "Creating shims as they either do not exist, or are outdated"
-            );
-
-            self.inventory.manifest.shim_version = SHIM_VERSION;
-            self.inventory.manifest.save()?;
-        }
-
-        let mut registry: ShimsMap = BTreeMap::default();
-        registry.insert(self.id.to_string(), Shim::default());
-
-        let mut to_create = vec![];
-
-        for shim in shims {
-            let mut shim_entry = Shim::default();
-
-            // Handle before and after args
-            if let Some(before_args) = shim.config.shim_before_args {
-                shim_entry.before_args = match before_args {
-                    StringOrVec::String(value) => shell_words::split(&value).into_diagnostic()?,
-                    StringOrVec::Vec(value) => value,
-                };
-            }
-
-            if let Some(after_args) = shim.config.shim_after_args {
-                shim_entry.after_args = match after_args {
-                    StringOrVec::String(value) => shell_words::split(&value).into_diagnostic()?,
-                    StringOrVec::Vec(value) => value,
-                };
-            }
-
-            if let Some(env_vars) = shim.config.shim_env_vars {
-                shim_entry.env_vars.extend(env_vars);
-            }
-
-            if !shim.primary {
-                shim_entry.parent = Some(self.id.to_string());
-
-                // Only use --alt when the secondary executable exists
-                if shim.config.exe_path.is_some() {
-                    shim_entry.alt_bin = Some(true);
-                }
-            }
-
-            // Create the shim file by copying the source bin
-            if force_create || find_only && !shim.path.exists() {
-                to_create.push(shim.path);
-            }
-
-            // Update the registry
-            registry.insert(shim.name.clone(), shim_entry);
-        }
-
-        // Only lock the directory and create shims if necessary
-        if !to_create.is_empty() {
-            let _lock = fs::lock_directory(&self.proto.store.shims_dir)?;
-
-            for shim_path in to_create {
-                self.proto.store.create_shim(&shim_path)?;
-
-                debug!(
-                    tool = self.id.as_str(),
-                    shim = ?shim_path,
-                    shim_version = SHIM_VERSION,
-                    "Creating shim"
-                );
-            }
-
-            ShimRegistry::update(&self.proto, registry)?;
-        }
-
-        Ok(())
-    }
-
-    /// Symlink all primary and secondary binaries for the current tool.
-    #[instrument(skip(self))]
-    pub async fn symlink_bins(&mut self, force: bool) -> miette::Result<()> {
-        let bins = self.resolve_bin_locations().await?;
-
-        if bins.is_empty() {
-            return Ok(());
-        }
-
-        if force {
-            debug!(
-                tool = self.id.as_str(),
-                bins_dir = ?self.proto.store.bin_dir,
-                "Creating symlinks to the original tool executables"
-            );
-        }
-
-        let tool_dir = self.get_product_dir();
-        let mut to_create = vec![];
-
-        for bin in bins {
-            let input_path = tool_dir.join(
-                bin.config
-                    .exe_link_path
-                    .as_ref()
-                    .or(bin.config.exe_path.as_ref())
-                    .unwrap(),
-            );
-            let output_path = bin.path;
-
-            if !input_path.exists() {
-                warn!(
-                    tool = self.id.as_str(),
-                    source = ?input_path,
-                    target = ?output_path,
-                    "Unable to symlink binary, source file does not exist"
-                );
-
-                continue;
-            }
-
-            if !force && output_path.exists() {
-                continue;
-            }
-
-            to_create.push((input_path, output_path));
-        }
-
-        // Only lock the directory and create bins if necessary
-        if !to_create.is_empty() {
-            let _lock = fs::lock_directory(&self.proto.store.bin_dir)?;
-
-            for (input_path, output_path) in to_create {
-                debug!(
-                    tool = self.id.as_str(),
-                    source = ?input_path,
-                    target = ?output_path,
-                    "Creating binary symlink"
-                );
-
-                self.proto.store.unlink_bin(&output_path)?;
-                self.proto.store.link_bin(&output_path, &input_path)?;
-            }
         }
 
         Ok(())
